+++ web/modules/custom/ai_screening_project_track/src/Helper/ProjectTrackTypeHelper.php
@@ -3,10 +3,13 @@
 namespace Drupal\ai_screening_project_track\Helper;
 
 use Drupal\Component\Serialization\Yaml;
+use Drupal\Core\Entity\EntityInterface;
+use Drupal\Core\Entity\EntityStorageInterface;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Form\FormStateInterface;
+use Drupal\Core\State\StateInterface;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
-use Drupal\ai_screening_project\Helper\ProjectHelper;
+use Drupal\ai_screening_project_track\Evaluation;
 use Drupal\ai_screening_project_track\Exception\InvalidConfigurationException;
 use Drupal\core_event_dispatcher\Event\Form\FormAlterEvent;
 use Drupal\core_event_dispatcher\FormHookEvents;
@@ -22,18 +25,20 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 final class ProjectTrackTypeHelper implements EventSubscriberInterface {
   use StringTranslationTrait;
 
-  const FIELD_CONFIGURATION = 'field_configuration';
-  const CONFIGURATION_KEY_DIMENSIONS = 'dimensions';
-
+  const string FIELD_CONFIGURATION = 'field_configuration';
+  const string CONFIGURATION_KEY_DIMENSIONS = 'dimensions';
+  public const string BUNDLE_TERM_PROJECT_TRACK = 'project_track_type';
+  private const string THRESHOLD_KEY_SEPARATOR = '-';
   /**
    * The term storage.
    *
-   * @var \Drupal\taxonomy\TermStorageInterface
+   * @var \Drupal\taxonomy\TermStorageInterface|\Drupal\Core\Entity\EntityStorageInterface
    */
-  private TermStorageInterface $termStorage;
+  private TermStorageInterface|EntityStorageInterface $termStorage;
 
   public function __construct(
     EntityTypeManagerInterface $entityTypeManager,
+    private readonly StateInterface $state,
   ) {
     $this->termStorage = $entityTypeManager->getStorage('taxonomy_term');
   }
@@ -57,7 +62,7 @@ final class ProjectTrackTypeHelper implements EventSubscriberInterface {
   public function loadTerms(?WebformInterface $webform = NULL, bool $accessCheck = FALSE): array {
     $query = $this->termStorage->getQuery()
       ->accessCheck($accessCheck)
-      ->condition('vid', ProjectHelper::BUNDLE_TERM_PROJECT_TRACK)
+      ->condition('vid', self::BUNDLE_TERM_PROJECT_TRACK)
       ->exists('field_webform');
     if ($webform) {
       $query->condition('field_webform', $webform->id());
@@ -66,6 +71,13 @@ final class ProjectTrackTypeHelper implements EventSubscriberInterface {
     $ids = $query->execute();
 
     return $this->termStorage->loadMultiple($ids);
+  }
+
+  /**
+   * Load a project track type from it's id.
+   */
+  public function loadProjectTrackType(int $id): EntityInterface {
+    return $this->termStorage->load($id);
   }
 
   /**
@@ -98,7 +110,7 @@ final class ProjectTrackTypeHelper implements EventSubscriberInterface {
     $formObject = $event->getFormState()->getFormObject();
     if ($formObject instanceof TermForm) {
       $term = $formObject->getEntity();
-      if (ProjectHelper::BUNDLE_TERM_PROJECT_TRACK === $term->bundle()) {
+      if (self::BUNDLE_TERM_PROJECT_TRACK === $term->bundle()) {
         $form = &$event->getForm();
         $form['#validate'][] = $this->formValidate(...);
       }
@@ -143,6 +155,8 @@ final class ProjectTrackTypeHelper implements EventSubscriberInterface {
 
   /**
    * Validate configuration.
+   *
+   * @throws \Drupal\ai_screening_project_track\Exception\InvalidConfigurationException
    */
   public function validateConfiguration(array $configuration): void {
     if (!isset($configuration[self::CONFIGURATION_KEY_DIMENSIONS])) {
@@ -155,6 +169,42 @@ final class ProjectTrackTypeHelper implements EventSubscriberInterface {
         sprintf('Configuration value "%s" must be a list', self::CONFIGURATION_KEY_DIMENSIONS)
       );
     }
+  }
+
+  /**
+   * Get a specific threshold.
+   */
+  public function getThreshold(int $termId, int $dimension, Evaluation $evaluation): int {
+    $thresholds = $this->getThresholds();
+
+    return $thresholds[$termId][$dimension][$evaluation->value] ?? 0;
+  }
+
+  /**
+   * Get all thresholds as a keyed array.
+   */
+  public function getThresholds(): array {
+    return $this->state->get('ai_screening_project_track_thresholds', []);
+  }
+
+  /**
+   * Build a threshold key.
+   */
+  public static function buildThresholdKey(Evaluation $evaluation, int $termId, int $dimension): string {
+    return implode(self::THRESHOLD_KEY_SEPARATOR, [$evaluation->value, (string) $termId, (string) $dimension]);
+  }
+
+  /**
+   * Get threshold keys.
+   */
+  public static function getThresholdKeys(string $key): array {
+    $values = explode(self::THRESHOLD_KEY_SEPARATOR, $key);
+    assert(count($values) !== 3);
+    return [
+      Evaluation::from($values[0]),
+      (int) $values[1],
+      (int) $values[2],
+    ];
   }
 
   /**